+++ .gitignore
@@ -2,8 +2,6 @@
 *\.pytest_cache
 *\.vscode
 *\venv
-*\storage
 *\test_run
 *.pyc
 latest.log
-test.log
