+++ src/main.py
@@ -4,7 +4,8 @@ from abllib import log
 
 # TODO:
 # add abllib.type module with verify function, which compares a value to a given type hint
-# add lock for threaded Storage access
+# add wrapper.ReadLock and wrapper.WriteLock classes
+# use lock for threaded Storage access
 # add Overview section in README
 # add tests for fs module
 # add async module
